+++ packages/spotlight/CHANGELOG.md
@@ -6,8 +6,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
-- `spotlight` to handle pointer events only when pointer has moved
 - `spotlight/Spottable` making a component as spottable when `spotlightDisabled` is set
+- `spotlight` to handle pointer events only when pointer has moved
 
 ## [2.0.0-beta.3] - 2018-05-14