+++ packages/spotlight/CHANGELOG.md
@@ -2,11 +2,24 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [2.0.0-beta.4] - 2018-05-21
+
+### Fixed
+
+- `spotlight/Spottable` to not make components spottable when `spotlightDisabled` is set
+
+## [2.0.0-beta.3] - 2018-05-14
+
+### Fixed
+
+- `spotlight` to retry setting focus when the window is activated
+- `spotlight` handling of 5-way events after the pointer hides
+
 ## [2.0.0-beta.2] - 2018-05-07
 
 ### Fixed
 
-- `spotlight/Spottable` to not set its focused state when disabled
+- `spotlight/Spottable` to not add a focused state when a component had already been set as disabled
 
 ## [2.0.0-beta.1] - 2018-04-29